Move template.gen.go outside the /pkg/template dir. Also separate the Sample template related generated code from out built in template related generated code.
