+++ cmd/server/BUILD
@@ -14,6 +14,7 @@ go_binary(
         "//cmd/server/cmd:go_default_library",
         "//cmd/shared:go_default_library",
         "//pkg/template:go_default_library",
+        "//template:go_default_library",
         "@com_github_spf13_cobra//:go_default_library",
     ],
 )
