+++ pkg/adapterManager/e2e_configuration_test.go
@@ -31,6 +31,7 @@ import (
 	"istio.io/mixer/pkg/expr"
 	"istio.io/mixer/pkg/pool"
 	"istio.io/mixer/pkg/template"
+	"istio.io/mixer/template/sample"
 	sample_report "istio.io/mixer/template/sample/report"
 )
 
@@ -174,7 +175,7 @@ func testConfigFlow(t *testing.T, declarativeSrvcCnfgFilePath string, declaredGl
 	}
 
 	cnfgMgr := config.NewManager(eval, adapterMgr.AspectValidatorFinder, adapterMgr.BuilderValidatorFinder, []adapter.GetBuilderInfoFn{GetFakeHndlrBuilderInfo},
-		adapterMgr.SupportedKinds, template.NewRepository(template.SupportedTmplInfo), store,
+		adapterMgr.SupportedKinds, template.NewRepository(sample.SupportedTmplInfo), store,
 		loopDelay,
 		identityAttribute, identityDomainAttribute)
 	cnfgMgr.Register(adapterMgr)
