+++ template/sample/BUILD
@@ -5,7 +5,7 @@ load("@io_bazel_rules_go//go:def.bzl", "go_library", "go_test")
 go_library(
     name = "go_default_library",
     srcs = [
-        "template.go",
+        "template.gen.go",
     ],
     visibility = ["//visibility:public"],
     deps = [
@@ -14,12 +14,7 @@ go_library(
         "//pkg/attribute:go_default_library",
         "//pkg/expr:go_default_library",
         "//pkg/status:go_default_library",
-        "//template/checknothing:go_default_library",
-        "//template/listentry:go_default_library",
-        "//template/logentry:go_default_library",
-        "//template/metric:go_default_library",
-        "//template/quota:go_default_library",
-        "//template/reportnothing:go_default_library",
+        "//pkg/template:go_default_library",
         "//template/sample/check:go_default_library",
         "//template/sample/quota:go_default_library",
         "//template/sample/report:go_default_library",
@@ -35,7 +30,7 @@ go_test(
     name = "go_default_test",
     size = "small",
     srcs = [
-        "template_test.go",
+        "template.gen_test.go",
     ],
     library = ":go_default_library",
     deps = [
