+++ template/sample/template.gen.go
@@ -14,7 +14,7 @@
 
 // THIS FILE IS AUTOMATICALLY GENERATED.
 
-package tmp
+package sample
 
 import (
 	"fmt"
@@ -32,35 +32,33 @@ import (
 	"istio.io/mixer/pkg/status"
 	"istio.io/mixer/pkg/template"
 
-	"istio.io/mixer/template/list"
+	"istio.io/mixer/template/sample/check"
 
-	"istio.io/mixer/template/log"
+	"istio.io/mixer/template/sample/quota"
 
-	"istio.io/mixer/template/metric"
-
-	"istio.io/mixer/template/quota"
+	"istio.io/mixer/template/sample/report"
 )
 
 var (
 	SupportedTmplInfo = map[string]template.Info{
 
-		istio_mixer_template_list.TemplateName: {
-			CtrCfg:    &istio_mixer_template_list.InstanceParam{},
+		istio_mixer_adapter_sample_check.TemplateName: {
+			CtrCfg:    &istio_mixer_adapter_sample_check.InstanceParam{},
 			Variety:   adptTmpl.TEMPLATE_VARIETY_CHECK,
-			BldrName:  "istio.io/mixer/template/list.ListHandlerBuilder",
-			HndlrName: "istio.io/mixer/template/list.ListHandler",
+			BldrName:  "istio.io/mixer/template/sample/check.SampleHandlerBuilder",
+			HndlrName: "istio.io/mixer/template/sample/check.SampleHandler",
 			SupportsTemplate: func(hndlrBuilder adapter.HandlerBuilder) bool {
-				_, ok := hndlrBuilder.(istio_mixer_template_list.ListHandlerBuilder)
+				_, ok := hndlrBuilder.(istio_mixer_adapter_sample_check.SampleHandlerBuilder)
 				return ok
 			},
 			HandlerSupportsTemplate: func(hndlr adapter.Handler) bool {
-				_, ok := hndlr.(istio_mixer_template_list.ListHandler)
+				_, ok := hndlr.(istio_mixer_adapter_sample_check.SampleHandler)
 				return ok
 			},
 			InferType: func(cp proto.Message, tEvalFn template.TypeEvalFn) (proto.Message, error) {
 				var err error = nil
-				cpb := cp.(*istio_mixer_template_list.InstanceParam)
-				infrdType := &istio_mixer_template_list.Type{}
+				cpb := cp.(*istio_mixer_adapter_sample_check.InstanceParam)
+				infrdType := &istio_mixer_adapter_sample_check.Type{}
 
 				if cpb.CheckExpression == "" {
 					return nil, fmt.Errorf("expression for field CheckExpression cannot be empty")
@@ -78,14 +76,14 @@ var (
 			},
 			ConfigureType: func(types map[string]proto.Message, builder *adapter.HandlerBuilder) error {
 				// Mixer framework should have ensured the type safety.
-				castedBuilder := (*builder).(istio_mixer_template_list.ListHandlerBuilder)
-				castedTypes := make(map[string]*istio_mixer_template_list.Type, len(types))
+				castedBuilder := (*builder).(istio_mixer_adapter_sample_check.SampleHandlerBuilder)
+				castedTypes := make(map[string]*istio_mixer_adapter_sample_check.Type, len(types))
 				for k, v := range types {
 					// Mixer framework should have ensured the type safety.
-					v1 := v.(*istio_mixer_template_list.Type)
+					v1 := v.(*istio_mixer_adapter_sample_check.Type)
 					castedTypes[k] = v1
 				}
-				return castedBuilder.ConfigureListHandler(castedTypes)
+				return castedBuilder.ConfigureSampleHandler(castedTypes)
 			},
 
 			ProcessCheck: func(insts map[string]proto.Message, attrs attribute.Bag, mapper expr.Evaluator,
@@ -93,10 +91,10 @@ var (
 				var found bool
 				var err error
 
-				var instances []*istio_mixer_template_list.Instance
-				castedInsts := make(map[string]*istio_mixer_template_list.InstanceParam, len(insts))
+				var instances []*istio_mixer_adapter_sample_check.Instance
+				castedInsts := make(map[string]*istio_mixer_adapter_sample_check.InstanceParam, len(insts))
 				for k, v := range insts {
-					v1 := v.(*istio_mixer_template_list.InstanceParam)
+					v1 := v.(*istio_mixer_adapter_sample_check.InstanceParam)
 					castedInsts[k] = v1
 				}
 				for name, md := range castedInsts {
@@ -107,7 +105,7 @@ var (
 						return status.WithError(err), adapter.CacheabilityInfo{}
 					}
 
-					instances = append(instances, &istio_mixer_template_list.Instance{
+					instances = append(instances, &istio_mixer_adapter_sample_check.Instance{
 						Name: name,
 
 						CheckExpression: CheckExpression.(string),
@@ -115,7 +113,7 @@ var (
 					_ = md
 				}
 				var cacheInfo adapter.CacheabilityInfo
-				if found, cacheInfo, err = handler.(istio_mixer_template_list.ListHandler).HandleList(instances); err != nil {
+				if found, cacheInfo, err = handler.(istio_mixer_adapter_sample_check.SampleHandler).HandleSample(instances); err != nil {
 					return status.WithError(err), adapter.CacheabilityInfo{}
 				}
 
@@ -129,23 +127,23 @@ var (
 			ProcessQuota:  nil,
 		},
 
-		istio_mixer_template_log.TemplateName: {
-			CtrCfg:    &istio_mixer_template_log.InstanceParam{},
-			Variety:   adptTmpl.TEMPLATE_VARIETY_REPORT,
-			BldrName:  "istio.io/mixer/template/log.LogHandlerBuilder",
-			HndlrName: "istio.io/mixer/template/log.LogHandler",
+		istio_mixer_adapter_sample_quota.TemplateName: {
+			CtrCfg:    &istio_mixer_adapter_sample_quota.InstanceParam{},
+			Variety:   adptTmpl.TEMPLATE_VARIETY_QUOTA,
+			BldrName:  "istio.io/mixer/template/sample/quota.QuotaHandlerBuilder",
+			HndlrName: "istio.io/mixer/template/sample/quota.QuotaHandler",
 			SupportsTemplate: func(hndlrBuilder adapter.HandlerBuilder) bool {
-				_, ok := hndlrBuilder.(istio_mixer_template_log.LogHandlerBuilder)
+				_, ok := hndlrBuilder.(istio_mixer_adapter_sample_quota.QuotaHandlerBuilder)
 				return ok
 			},
 			HandlerSupportsTemplate: func(hndlr adapter.Handler) bool {
-				_, ok := hndlr.(istio_mixer_template_log.LogHandler)
+				_, ok := hndlr.(istio_mixer_adapter_sample_quota.QuotaHandler)
 				return ok
 			},
 			InferType: func(cp proto.Message, tEvalFn template.TypeEvalFn) (proto.Message, error) {
 				var err error = nil
-				cpb := cp.(*istio_mixer_template_log.InstanceParam)
-				infrdType := &istio_mixer_template_log.Type{}
+				cpb := cp.(*istio_mixer_adapter_sample_quota.InstanceParam)
+				infrdType := &istio_mixer_adapter_sample_quota.Type{}
 
 				infrdType.Dimensions = make(map[string]istio_mixer_v1_config_descriptor.ValueType, len(cpb.Dimensions))
 				for k, v := range cpb.Dimensions {
@@ -159,74 +157,71 @@ var (
 			},
 			ConfigureType: func(types map[string]proto.Message, builder *adapter.HandlerBuilder) error {
 				// Mixer framework should have ensured the type safety.
-				castedBuilder := (*builder).(istio_mixer_template_log.LogHandlerBuilder)
-				castedTypes := make(map[string]*istio_mixer_template_log.Type, len(types))
+				castedBuilder := (*builder).(istio_mixer_adapter_sample_quota.QuotaHandlerBuilder)
+				castedTypes := make(map[string]*istio_mixer_adapter_sample_quota.Type, len(types))
 				for k, v := range types {
 					// Mixer framework should have ensured the type safety.
-					v1 := v.(*istio_mixer_template_log.Type)
+					v1 := v.(*istio_mixer_adapter_sample_quota.Type)
 					castedTypes[k] = v1
 				}
-				return castedBuilder.ConfigureLogHandler(castedTypes)
+				return castedBuilder.ConfigureQuotaHandler(castedTypes)
 			},
 
-			ProcessReport: func(insts map[string]proto.Message, attrs attribute.Bag, mapper expr.Evaluator, handler adapter.Handler) rpc.Status {
-				result := &multierror.Error{}
-				var instances []*istio_mixer_template_log.Instance
+			ProcessQuota: func(quotaName string, inst proto.Message, attrs attribute.Bag, mapper expr.Evaluator, handler adapter.Handler,
+				qma adapter.QuotaRequestArgs) (rpc.Status, adapter.CacheabilityInfo, adapter.QuotaResult) {
+				castedInst := inst.(*istio_mixer_adapter_sample_quota.InstanceParam)
 
-				castedInsts := make(map[string]*istio_mixer_template_log.InstanceParam, len(insts))
-				for k, v := range insts {
-					v1 := v.(*istio_mixer_template_log.InstanceParam)
-					castedInsts[k] = v1
-				}
-				for name, md := range castedInsts {
+				Dimensions, err := template.EvalAll(castedInst.Dimensions, attrs, mapper)
 
-					Dimensions, err := template.EvalAll(md.Dimensions, attrs, mapper)
-
-					if err != nil {
-						result = multierror.Append(result, fmt.Errorf("failed to eval Dimensions for instance '%s': %v", name, err))
-						continue
-					}
-
-					instances = append(instances, &istio_mixer_template_log.Instance{
-						Name: name,
-
-						Dimensions: Dimensions,
-					})
-					_ = md
-				}
-
-				if err := handler.(istio_mixer_template_log.LogHandler).HandleLog(instances); err != nil {
-					result = multierror.Append(result, fmt.Errorf("failed to report all values: %v", err))
-				}
-
-				err := result.ErrorOrNil()
 				if err != nil {
-					return status.WithError(err)
+					msg := fmt.Sprintf("failed to eval Dimensions for instance '%s': %v", quotaName, err)
+					glog.Error(msg)
+					return status.WithInvalidArgument(msg), adapter.CacheabilityInfo{}, adapter.QuotaResult{}
 				}
 
-				return status.OK
+				instance := &istio_mixer_adapter_sample_quota.Instance{
+					Name: quotaName,
+
+					Dimensions: Dimensions,
+				}
+
+				var qr adapter.QuotaResult
+				var cacheInfo adapter.CacheabilityInfo
+				if qr, cacheInfo, err = handler.(istio_mixer_adapter_sample_quota.QuotaHandler).HandleQuota(instance, qma); err != nil {
+					glog.Errorf("Quota allocation failed: %v", err)
+					return status.WithError(err), adapter.CacheabilityInfo{}, adapter.QuotaResult{}
+				}
+				if qr.Amount == 0 {
+					msg := fmt.Sprintf("Unable to allocate %v units from quota %s", qma.QuotaAmount, quotaName)
+					glog.Warning(msg)
+					return status.WithResourceExhausted(msg), adapter.CacheabilityInfo{}, adapter.QuotaResult{}
+				}
+				if glog.V(2) {
+					glog.Infof("Allocated %v units from quota %s", qma.QuotaAmount, quotaName)
+				}
+				return status.OK, cacheInfo, qr
 			},
-			ProcessCheck: nil,
-			ProcessQuota: nil,
+			ProcessReport: nil,
+			ProcessCheck:  nil,
 		},
 
-		istio_mixer_template_metric.TemplateName: {
-			CtrCfg:    &istio_mixer_template_metric.InstanceParam{},
+		istio_mixer_adapter_sample_report.TemplateName: {
+			CtrCfg:    &istio_mixer_adapter_sample_report.InstanceParam{},
 			Variety:   adptTmpl.TEMPLATE_VARIETY_REPORT,
-			BldrName:  "istio.io/mixer/template/metric.MetricHandlerBuilder",
-			HndlrName: "istio.io/mixer/template/metric.MetricHandler",
+			BldrName:  "istio.io/mixer/template/sample/report.SampleHandlerBuilder",
+			HndlrName: "istio.io/mixer/template/sample/report.SampleHandler",
 			SupportsTemplate: func(hndlrBuilder adapter.HandlerBuilder) bool {
-				_, ok := hndlrBuilder.(istio_mixer_template_metric.MetricHandlerBuilder)
+				_, ok := hndlrBuilder.(istio_mixer_adapter_sample_report.SampleHandlerBuilder)
 				return ok
 			},
 			HandlerSupportsTemplate: func(hndlr adapter.Handler) bool {
-				_, ok := hndlr.(istio_mixer_template_metric.MetricHandler)
+				_, ok := hndlr.(istio_mixer_adapter_sample_report.SampleHandler)
 				return ok
 			},
 			InferType: func(cp proto.Message, tEvalFn template.TypeEvalFn) (proto.Message, error) {
 				var err error = nil
-				cpb := cp.(*istio_mixer_template_metric.InstanceParam)
-				infrdType := &istio_mixer_template_metric.Type{}
+				cpb := cp.(*istio_mixer_adapter_sample_report.InstanceParam)
+				infrdType := &istio_mixer_adapter_sample_report.Type{}
 
 				if cpb.Value == "" {
 					return nil, fmt.Errorf("expression for field Value cannot be empty")
@@ -287,36 +282,28 @@ var (
 					return nil, fmt.Errorf("error type checking for field StringPrimitive: Evaluated expression type %v want %v", t, istio_mixer_v1_config_descriptor.STRING)
 				}
 
-				if cpb.AnotherValueType == "" {
-					return nil, fmt.Errorf("expression for field AnotherValueType cannot be empty")
-				}
-
-				if infrdType.AnotherValueType, err = tEvalFn(cpb.AnotherValueType); err != nil {
-					return nil, err
-				}
-
 				_ = cpb
 				return infrdType, err
 			},
 			ConfigureType: func(types map[string]proto.Message, builder *adapter.HandlerBuilder) error {
 				// Mixer framework should have ensured the type safety.
-				castedBuilder := (*builder).(istio_mixer_template_metric.MetricHandlerBuilder)
-				castedTypes := make(map[string]*istio_mixer_template_metric.Type, len(types))
+				castedBuilder := (*builder).(istio_mixer_adapter_sample_report.SampleHandlerBuilder)
+				castedTypes := make(map[string]*istio_mixer_adapter_sample_report.Type, len(types))
 				for k, v := range types {
 					// Mixer framework should have ensured the type safety.
-					v1 := v.(*istio_mixer_template_metric.Type)
+					v1 := v.(*istio_mixer_adapter_sample_report.Type)
 					castedTypes[k] = v1
 				}
-				return castedBuilder.ConfigureMetricHandler(castedTypes)
+				return castedBuilder.ConfigureSampleHandler(castedTypes)
 			},
 
 			ProcessReport: func(insts map[string]proto.Message, attrs attribute.Bag, mapper expr.Evaluator, handler adapter.Handler) rpc.Status {
 				result := &multierror.Error{}
-				var instances []*istio_mixer_template_metric.Instance
+				var instances []*istio_mixer_adapter_sample_report.Instance
 
-				castedInsts := make(map[string]*istio_mixer_template_metric.InstanceParam, len(insts))
+				castedInsts := make(map[string]*istio_mixer_adapter_sample_report.InstanceParam, len(insts))
 				for k, v := range insts {
-					v1 := v.(*istio_mixer_template_metric.InstanceParam)
+					v1 := v.(*istio_mixer_adapter_sample_report.InstanceParam)
 					castedInsts[k] = v1
 				}
 				for name, md := range castedInsts {
@@ -363,14 +350,7 @@ var (
 						continue
 					}
 
-					AnotherValueType, err := mapper.Eval(md.AnotherValueType, attrs)
-
-					if err != nil {
-						result = multierror.Append(result, fmt.Errorf("failed to eval AnotherValueType for instance '%s': %v", name, err))
-						continue
-					}
-
-					instances = append(instances, &istio_mixer_template_metric.Instance{
+					instances = append(instances, &istio_mixer_adapter_sample_report.Instance{
 						Name: name,
 
 						Value: Value,
@@ -384,94 +364,11 @@ var (
 						DoublePrimitive: DoublePrimitive.(float64),
 
 						StringPrimitive: StringPrimitive.(string),
-
-						AnotherValueType: AnotherValueType,
 					})
 					_ = md
 				}
 
-				if err := handler.(istio_mixer_template_metric.MetricHandler).HandleMetric(instances); err != nil {
-					result = multierror.Append(result, fmt.Errorf("failed to report all values: %v", err))
-				}
-
-				err := result.ErrorOrNil()
-				if err != nil {
-					return status.WithError(err)
-				}
-
-				return status.OK
-			},
-			ProcessCheck: nil,
-			ProcessQuota: nil,
-		},
-
-		istio_mixer_template_quota.TemplateName: {
-			CtrCfg:    &istio_mixer_template_quota.InstanceParam{},
-			Variety:   adptTmpl.TEMPLATE_VARIETY_REPORT,
-			BldrName:  "istio.io/mixer/template/quota.QuotaHandlerBuilder",
-			HndlrName: "istio.io/mixer/template/quota.QuotaHandler",
-			SupportsTemplate: func(hndlrBuilder adapter.HandlerBuilder) bool {
-				_, ok := hndlrBuilder.(istio_mixer_template_quota.QuotaHandlerBuilder)
-				return ok
-			},
-			HandlerSupportsTemplate: func(hndlr adapter.Handler) bool {
-				_, ok := hndlr.(istio_mixer_template_quota.QuotaHandler)
-				return ok
-			},
-			InferType: func(cp proto.Message, tEvalFn template.TypeEvalFn) (proto.Message, error) {
-				var err error = nil
-				cpb := cp.(*istio_mixer_template_quota.InstanceParam)
-				infrdType := &istio_mixer_template_quota.Type{}
-
-				infrdType.Dimensions = make(map[string]istio_mixer_v1_config_descriptor.ValueType, len(cpb.Dimensions))
-				for k, v := range cpb.Dimensions {
-					if infrdType.Dimensions[k], err = tEvalFn(v); err != nil {
-						return nil, err
-					}
-				}
-
-				_ = cpb
-				return infrdType, err
-			},
-			ConfigureType: func(types map[string]proto.Message, builder *adapter.HandlerBuilder) error {
-				// Mixer framework should have ensured the type safety.
-				castedBuilder := (*builder).(istio_mixer_template_quota.QuotaHandlerBuilder)
-				castedTypes := make(map[string]*istio_mixer_template_quota.Type, len(types))
-				for k, v := range types {
-					// Mixer framework should have ensured the type safety.
-					v1 := v.(*istio_mixer_template_quota.Type)
-					castedTypes[k] = v1
-				}
-				return castedBuilder.ConfigureQuotaHandler(castedTypes)
-			},
-
-			ProcessReport: func(insts map[string]proto.Message, attrs attribute.Bag, mapper expr.Evaluator, handler adapter.Handler) rpc.Status {
-				result := &multierror.Error{}
-				var instances []*istio_mixer_template_quota.Instance
-
-				castedInsts := make(map[string]*istio_mixer_template_quota.InstanceParam, len(insts))
-				for k, v := range insts {
-					v1 := v.(*istio_mixer_template_quota.InstanceParam)
-					castedInsts[k] = v1
-				}
-				for name, md := range castedInsts {
-
-					Dimensions, err := template.EvalAll(md.Dimensions, attrs, mapper)
-
-					if err != nil {
-						result = multierror.Append(result, fmt.Errorf("failed to eval Dimensions for instance '%s': %v", name, err))
-						continue
-					}
-
-					instances = append(instances, &istio_mixer_template_quota.Instance{
-						Name: name,
-
-						Dimensions: Dimensions,
-					})
-					_ = md
-				}
-
-				if err := handler.(istio_mixer_template_quota.QuotaHandler).HandleQuota(instances); err != nil {
+				if err := handler.(istio_mixer_adapter_sample_report.SampleHandler).HandleSample(instances); err != nil {
 					result = multierror.Append(result, fmt.Errorf("failed to report all values: %v", err))
 				}
