+++ template/template.gen.go
@@ -30,6 +30,7 @@ import (
 	"istio.io/mixer/pkg/attribute"
 	"istio.io/mixer/pkg/expr"
 	"istio.io/mixer/pkg/status"
+	"istio.io/mixer/pkg/template"
 
 	"istio.io/mixer/template/checknothing"
 
@@ -42,16 +43,10 @@ import (
 	"istio.io/mixer/template/quota"
 
 	"istio.io/mixer/template/reportnothing"
-
-	"istio.io/mixer/template/sample/check"
-
-	"istio.io/mixer/template/sample/quota"
-
-	"istio.io/mixer/template/sample/report"
 )
 
 var (
-	SupportedTmplInfo = map[string]Info{
+	SupportedTmplInfo = map[string]template.Info{
 
 		checknothing.TemplateName: {
 			CtrCfg:    &checknothing.InstanceParam{},
@@ -66,7 +61,7 @@ var (
 				_, ok := hndlr.(checknothing.CheckNothingHandler)
 				return ok
 			},
-			InferType: func(cp proto.Message, tEvalFn TypeEvalFn) (proto.Message, error) {
+			InferType: func(cp proto.Message, tEvalFn template.TypeEvalFn) (proto.Message, error) {
 				var err error = nil
 				cpb := cp.(*checknothing.InstanceParam)
 				infrdType := &checknothing.Type{}
@@ -132,7 +127,7 @@ var (
 				_, ok := hndlr.(listentry.ListEntryHandler)
 				return ok
 			},
-			InferType: func(cp proto.Message, tEvalFn TypeEvalFn) (proto.Message, error) {
+			InferType: func(cp proto.Message, tEvalFn template.TypeEvalFn) (proto.Message, error) {
 				var err error = nil
 				cpb := cp.(*listentry.InstanceParam)
 				infrdType := &listentry.Type{}
@@ -217,7 +212,7 @@ var (
 				_, ok := hndlr.(logentry.LogEntryHandler)
 				return ok
 			},
-			InferType: func(cp proto.Message, tEvalFn TypeEvalFn) (proto.Message, error) {
+			InferType: func(cp proto.Message, tEvalFn template.TypeEvalFn) (proto.Message, error) {
 				var err error = nil
 				cpb := cp.(*logentry.InstanceParam)
 				infrdType := &logentry.Type{}
@@ -266,7 +261,7 @@ var (
 				}
 				for name, md := range castedInsts {
 
-					Labels, err := evalAll(md.Labels, attrs, mapper)
+					Labels, err := template.EvalAll(md.Labels, attrs, mapper)
 
 					if err != nil {
 						result = multierror.Append(result, fmt.Errorf("failed to eval Labels for instance '%s': %v", name, err))
@@ -318,7 +313,7 @@ var (
 				_, ok := hndlr.(metric.MetricHandler)
 				return ok
 			},
-			InferType: func(cp proto.Message, tEvalFn TypeEvalFn) (proto.Message, error) {
+			InferType: func(cp proto.Message, tEvalFn template.TypeEvalFn) (proto.Message, error) {
 				var err error = nil
 				cpb := cp.(*metric.InstanceParam)
 				infrdType := &metric.Type{}
@@ -371,7 +366,7 @@ var (
 						continue
 					}
 
-					Dimensions, err := evalAll(md.Dimensions, attrs, mapper)
+					Dimensions, err := template.EvalAll(md.Dimensions, attrs, mapper)
 
 					if err != nil {
 						result = multierror.Append(result, fmt.Errorf("failed to eval Dimensions for instance '%s': %v", name, err))
@@ -416,7 +411,7 @@ var (
 				_, ok := hndlr.(quota.QuotaHandler)
 				return ok
 			},
-			InferType: func(cp proto.Message, tEvalFn TypeEvalFn) (proto.Message, error) {
+			InferType: func(cp proto.Message, tEvalFn template.TypeEvalFn) (proto.Message, error) {
 				var err error = nil
 				cpb := cp.(*quota.InstanceParam)
 				infrdType := &quota.Type{}
@@ -447,7 +442,7 @@ var (
 				qma adapter.QuotaRequestArgs) (rpc.Status, adapter.CacheabilityInfo, adapter.QuotaResult) {
 				castedInst := inst.(*quota.InstanceParam)
 
-				Dimensions, err := evalAll(castedInst.Dimensions, attrs, mapper)
+				Dimensions, err := template.EvalAll(castedInst.Dimensions, attrs, mapper)
 
 				if err != nil {
 					msg := fmt.Sprintf("failed to eval Dimensions for instance '%s': %v", quotaName, err)
@@ -494,7 +489,7 @@ var (
 				_, ok := hndlr.(reportnothing.ReportNothingHandler)
 				return ok
 			},
-			InferType: func(cp proto.Message, tEvalFn TypeEvalFn) (proto.Message, error) {
+			InferType: func(cp proto.Message, tEvalFn template.TypeEvalFn) (proto.Message, error) {
 				var err error = nil
 				cpb := cp.(*reportnothing.InstanceParam)
 				infrdType := &reportnothing.Type{}
@@ -532,347 +527,6 @@ var (
 				}
 
 				if err := handler.(reportnothing.ReportNothingHandler).HandleReportNothing(instances); err != nil {
-					result = multierror.Append(result, fmt.Errorf("failed to report all values: %v", err))
-				}
-
-				err := result.ErrorOrNil()
-				if err != nil {
-					return status.WithError(err)
-				}
-
-				return status.OK
-			},
-			ProcessCheck: nil,
-			ProcessQuota: nil,
-		},
-
-		istio_mixer_adapter_sample_check.TemplateName: {
-			CtrCfg:    &istio_mixer_adapter_sample_check.InstanceParam{},
-			Variety:   adptTmpl.TEMPLATE_VARIETY_CHECK,
-			BldrName:  "istio.io/mixer/template/sample/check.SampleHandlerBuilder",
-			HndlrName: "istio.io/mixer/template/sample/check.SampleHandler",
-			SupportsTemplate: func(hndlrBuilder adapter.HandlerBuilder) bool {
-				_, ok := hndlrBuilder.(istio_mixer_adapter_sample_check.SampleHandlerBuilder)
-				return ok
-			},
-			HandlerSupportsTemplate: func(hndlr adapter.Handler) bool {
-				_, ok := hndlr.(istio_mixer_adapter_sample_check.SampleHandler)
-				return ok
-			},
-			InferType: func(cp proto.Message, tEvalFn TypeEvalFn) (proto.Message, error) {
-				var err error = nil
-				cpb := cp.(*istio_mixer_adapter_sample_check.InstanceParam)
-				infrdType := &istio_mixer_adapter_sample_check.Type{}
-
-				if cpb.CheckExpression == "" {
-					return nil, fmt.Errorf("expression for field CheckExpression cannot be empty")
-				}
-
-				if t, e := tEvalFn(cpb.CheckExpression); e != nil || t != istio_mixer_v1_config_descriptor.STRING {
-					if e != nil {
-						return nil, fmt.Errorf("failed to evaluate expression for field CheckExpression: %v", e)
-					}
-					return nil, fmt.Errorf("error type checking for field CheckExpression: Evaluated expression type %v want %v", t, istio_mixer_v1_config_descriptor.STRING)
-				}
-
-				_ = cpb
-				return infrdType, err
-			},
-			ConfigureType: func(types map[string]proto.Message, builder *adapter.HandlerBuilder) error {
-				// Mixer framework should have ensured the type safety.
-				castedBuilder := (*builder).(istio_mixer_adapter_sample_check.SampleHandlerBuilder)
-				castedTypes := make(map[string]*istio_mixer_adapter_sample_check.Type, len(types))
-				for k, v := range types {
-					// Mixer framework should have ensured the type safety.
-					v1 := v.(*istio_mixer_adapter_sample_check.Type)
-					castedTypes[k] = v1
-				}
-				return castedBuilder.ConfigureSampleHandler(castedTypes)
-			},
-
-			ProcessCheck: func(insts map[string]proto.Message, attrs attribute.Bag, mapper expr.Evaluator,
-				handler adapter.Handler) (rpc.Status, adapter.CacheabilityInfo) {
-				var found bool
-				var err error
-
-				var instances []*istio_mixer_adapter_sample_check.Instance
-				castedInsts := make(map[string]*istio_mixer_adapter_sample_check.InstanceParam, len(insts))
-				for k, v := range insts {
-					v1 := v.(*istio_mixer_adapter_sample_check.InstanceParam)
-					castedInsts[k] = v1
-				}
-				for name, md := range castedInsts {
-
-					CheckExpression, err := mapper.Eval(md.CheckExpression, attrs)
-
-					if err != nil {
-						return status.WithError(err), adapter.CacheabilityInfo{}
-					}
-
-					instances = append(instances, &istio_mixer_adapter_sample_check.Instance{
-						Name: name,
-
-						CheckExpression: CheckExpression.(string),
-					})
-					_ = md
-				}
-				var cacheInfo adapter.CacheabilityInfo
-				if found, cacheInfo, err = handler.(istio_mixer_adapter_sample_check.SampleHandler).HandleSample(instances); err != nil {
-					return status.WithError(err), adapter.CacheabilityInfo{}
-				}
-
-				if found {
-					return status.OK, cacheInfo
-				}
-
-				return status.WithPermissionDenied(fmt.Sprintf("%s rejected", instances)), adapter.CacheabilityInfo{}
-			},
-			ProcessReport: nil,
-			ProcessQuota:  nil,
-		},
-
-		istio_mixer_adapter_sample_quota.TemplateName: {
-			CtrCfg:    &istio_mixer_adapter_sample_quota.InstanceParam{},
-			Variety:   adptTmpl.TEMPLATE_VARIETY_QUOTA,
-			BldrName:  "istio.io/mixer/template/sample/quota.QuotaHandlerBuilder",
-			HndlrName: "istio.io/mixer/template/sample/quota.QuotaHandler",
-			SupportsTemplate: func(hndlrBuilder adapter.HandlerBuilder) bool {
-				_, ok := hndlrBuilder.(istio_mixer_adapter_sample_quota.QuotaHandlerBuilder)
-				return ok
-			},
-			HandlerSupportsTemplate: func(hndlr adapter.Handler) bool {
-				_, ok := hndlr.(istio_mixer_adapter_sample_quota.QuotaHandler)
-				return ok
-			},
-			InferType: func(cp proto.Message, tEvalFn TypeEvalFn) (proto.Message, error) {
-				var err error = nil
-				cpb := cp.(*istio_mixer_adapter_sample_quota.InstanceParam)
-				infrdType := &istio_mixer_adapter_sample_quota.Type{}
-
-				infrdType.Dimensions = make(map[string]istio_mixer_v1_config_descriptor.ValueType, len(cpb.Dimensions))
-				for k, v := range cpb.Dimensions {
-					if infrdType.Dimensions[k], err = tEvalFn(v); err != nil {
-						return nil, err
-					}
-				}
-
-				_ = cpb
-				return infrdType, err
-			},
-			ConfigureType: func(types map[string]proto.Message, builder *adapter.HandlerBuilder) error {
-				// Mixer framework should have ensured the type safety.
-				castedBuilder := (*builder).(istio_mixer_adapter_sample_quota.QuotaHandlerBuilder)
-				castedTypes := make(map[string]*istio_mixer_adapter_sample_quota.Type, len(types))
-				for k, v := range types {
-					// Mixer framework should have ensured the type safety.
-					v1 := v.(*istio_mixer_adapter_sample_quota.Type)
-					castedTypes[k] = v1
-				}
-				return castedBuilder.ConfigureQuotaHandler(castedTypes)
-			},
-
-			ProcessQuota: func(quotaName string, inst proto.Message, attrs attribute.Bag, mapper expr.Evaluator, handler adapter.Handler,
-				qma adapter.QuotaRequestArgs) (rpc.Status, adapter.CacheabilityInfo, adapter.QuotaResult) {
-				castedInst := inst.(*istio_mixer_adapter_sample_quota.InstanceParam)
-
-				Dimensions, err := evalAll(castedInst.Dimensions, attrs, mapper)
-
-				if err != nil {
-					msg := fmt.Sprintf("failed to eval Dimensions for instance '%s': %v", quotaName, err)
-					glog.Error(msg)
-					return status.WithInvalidArgument(msg), adapter.CacheabilityInfo{}, adapter.QuotaResult{}
-				}
-
-				instance := &istio_mixer_adapter_sample_quota.Instance{
-					Name: quotaName,
-
-					Dimensions: Dimensions,
-				}
-
-				var qr adapter.QuotaResult
-				var cacheInfo adapter.CacheabilityInfo
-				if qr, cacheInfo, err = handler.(istio_mixer_adapter_sample_quota.QuotaHandler).HandleQuota(instance, qma); err != nil {
-					glog.Errorf("Quota allocation failed: %v", err)
-					return status.WithError(err), adapter.CacheabilityInfo{}, adapter.QuotaResult{}
-				}
-				if qr.Amount == 0 {
-					msg := fmt.Sprintf("Unable to allocate %v units from quota %s", qma.QuotaAmount, quotaName)
-					glog.Warning(msg)
-					return status.WithResourceExhausted(msg), adapter.CacheabilityInfo{}, adapter.QuotaResult{}
-				}
-				if glog.V(2) {
-					glog.Infof("Allocated %v units from quota %s", qma.QuotaAmount, quotaName)
-				}
-				return status.OK, cacheInfo, qr
-			},
-			ProcessReport: nil,
-			ProcessCheck:  nil,
-		},
-
-		istio_mixer_adapter_sample_report.TemplateName: {
-			CtrCfg:    &istio_mixer_adapter_sample_report.InstanceParam{},
-			Variety:   adptTmpl.TEMPLATE_VARIETY_REPORT,
-			BldrName:  "istio.io/mixer/template/sample/report.SampleHandlerBuilder",
-			HndlrName: "istio.io/mixer/template/sample/report.SampleHandler",
-			SupportsTemplate: func(hndlrBuilder adapter.HandlerBuilder) bool {
-				_, ok := hndlrBuilder.(istio_mixer_adapter_sample_report.SampleHandlerBuilder)
-				return ok
-			},
-			HandlerSupportsTemplate: func(hndlr adapter.Handler) bool {
-				_, ok := hndlr.(istio_mixer_adapter_sample_report.SampleHandler)
-				return ok
-			},
-			InferType: func(cp proto.Message, tEvalFn TypeEvalFn) (proto.Message, error) {
-				var err error = nil
-				cpb := cp.(*istio_mixer_adapter_sample_report.InstanceParam)
-				infrdType := &istio_mixer_adapter_sample_report.Type{}
-
-				if cpb.Value == "" {
-					return nil, fmt.Errorf("expression for field Value cannot be empty")
-				}
-
-				if infrdType.Value, err = tEvalFn(cpb.Value); err != nil {
-					return nil, err
-				}
-
-				infrdType.Dimensions = make(map[string]istio_mixer_v1_config_descriptor.ValueType, len(cpb.Dimensions))
-				for k, v := range cpb.Dimensions {
-					if infrdType.Dimensions[k], err = tEvalFn(v); err != nil {
-						return nil, err
-					}
-				}
-
-				if cpb.Int64Primitive == "" {
-					return nil, fmt.Errorf("expression for field Int64Primitive cannot be empty")
-				}
-
-				if t, e := tEvalFn(cpb.Int64Primitive); e != nil || t != istio_mixer_v1_config_descriptor.INT64 {
-					if e != nil {
-						return nil, fmt.Errorf("failed to evaluate expression for field Int64Primitive: %v", e)
-					}
-					return nil, fmt.Errorf("error type checking for field Int64Primitive: Evaluated expression type %v want %v", t, istio_mixer_v1_config_descriptor.INT64)
-				}
-
-				if cpb.BoolPrimitive == "" {
-					return nil, fmt.Errorf("expression for field BoolPrimitive cannot be empty")
-				}
-
-				if t, e := tEvalFn(cpb.BoolPrimitive); e != nil || t != istio_mixer_v1_config_descriptor.BOOL {
-					if e != nil {
-						return nil, fmt.Errorf("failed to evaluate expression for field BoolPrimitive: %v", e)
-					}
-					return nil, fmt.Errorf("error type checking for field BoolPrimitive: Evaluated expression type %v want %v", t, istio_mixer_v1_config_descriptor.BOOL)
-				}
-
-				if cpb.DoublePrimitive == "" {
-					return nil, fmt.Errorf("expression for field DoublePrimitive cannot be empty")
-				}
-
-				if t, e := tEvalFn(cpb.DoublePrimitive); e != nil || t != istio_mixer_v1_config_descriptor.DOUBLE {
-					if e != nil {
-						return nil, fmt.Errorf("failed to evaluate expression for field DoublePrimitive: %v", e)
-					}
-					return nil, fmt.Errorf("error type checking for field DoublePrimitive: Evaluated expression type %v want %v", t, istio_mixer_v1_config_descriptor.DOUBLE)
-				}
-
-				if cpb.StringPrimitive == "" {
-					return nil, fmt.Errorf("expression for field StringPrimitive cannot be empty")
-				}
-
-				if t, e := tEvalFn(cpb.StringPrimitive); e != nil || t != istio_mixer_v1_config_descriptor.STRING {
-					if e != nil {
-						return nil, fmt.Errorf("failed to evaluate expression for field StringPrimitive: %v", e)
-					}
-					return nil, fmt.Errorf("error type checking for field StringPrimitive: Evaluated expression type %v want %v", t, istio_mixer_v1_config_descriptor.STRING)
-				}
-
-				_ = cpb
-				return infrdType, err
-			},
-			ConfigureType: func(types map[string]proto.Message, builder *adapter.HandlerBuilder) error {
-				// Mixer framework should have ensured the type safety.
-				castedBuilder := (*builder).(istio_mixer_adapter_sample_report.SampleHandlerBuilder)
-				castedTypes := make(map[string]*istio_mixer_adapter_sample_report.Type, len(types))
-				for k, v := range types {
-					// Mixer framework should have ensured the type safety.
-					v1 := v.(*istio_mixer_adapter_sample_report.Type)
-					castedTypes[k] = v1
-				}
-				return castedBuilder.ConfigureSampleHandler(castedTypes)
-			},
-
-			ProcessReport: func(insts map[string]proto.Message, attrs attribute.Bag, mapper expr.Evaluator, handler adapter.Handler) rpc.Status {
-				result := &multierror.Error{}
-				var instances []*istio_mixer_adapter_sample_report.Instance
-
-				castedInsts := make(map[string]*istio_mixer_adapter_sample_report.InstanceParam, len(insts))
-				for k, v := range insts {
-					v1 := v.(*istio_mixer_adapter_sample_report.InstanceParam)
-					castedInsts[k] = v1
-				}
-				for name, md := range castedInsts {
-
-					Value, err := mapper.Eval(md.Value, attrs)
-
-					if err != nil {
-						result = multierror.Append(result, fmt.Errorf("failed to eval Value for instance '%s': %v", name, err))
-						continue
-					}
-
-					Dimensions, err := evalAll(md.Dimensions, attrs, mapper)
-
-					if err != nil {
-						result = multierror.Append(result, fmt.Errorf("failed to eval Dimensions for instance '%s': %v", name, err))
-						continue
-					}
-
-					Int64Primitive, err := mapper.Eval(md.Int64Primitive, attrs)
-
-					if err != nil {
-						result = multierror.Append(result, fmt.Errorf("failed to eval Int64Primitive for instance '%s': %v", name, err))
-						continue
-					}
-
-					BoolPrimitive, err := mapper.Eval(md.BoolPrimitive, attrs)
-
-					if err != nil {
-						result = multierror.Append(result, fmt.Errorf("failed to eval BoolPrimitive for instance '%s': %v", name, err))
-						continue
-					}
-
-					DoublePrimitive, err := mapper.Eval(md.DoublePrimitive, attrs)
-
-					if err != nil {
-						result = multierror.Append(result, fmt.Errorf("failed to eval DoublePrimitive for instance '%s': %v", name, err))
-						continue
-					}
-
-					StringPrimitive, err := mapper.Eval(md.StringPrimitive, attrs)
-
-					if err != nil {
-						result = multierror.Append(result, fmt.Errorf("failed to eval StringPrimitive for instance '%s': %v", name, err))
-						continue
-					}
-
-					instances = append(instances, &istio_mixer_adapter_sample_report.Instance{
-						Name: name,
-
-						Value: Value,
-
-						Dimensions: Dimensions,
-
-						Int64Primitive: Int64Primitive.(int64),
-
-						BoolPrimitive: BoolPrimitive.(bool),
-
-						DoublePrimitive: DoublePrimitive.(float64),
-
-						StringPrimitive: StringPrimitive.(string),
-					})
-					_ = md
-				}
-
-				if err := handler.(istio_mixer_adapter_sample_report.SampleHandler).HandleSample(instances); err != nil {
 					result = multierror.Append(result, fmt.Errorf("failed to report all values: %v", err))
 				}
