+++ tools/codegen/pkg/bootstrapgen/template/bootstrapTemplate.go
@@ -45,6 +45,7 @@ import (
 	"istio.io/mixer/pkg/status"
 	"istio.io/mixer/pkg/adapter"
 	"istio.io/api/mixer/v1/config/descriptor"
+	"istio.io/mixer/pkg/template"
 	adptTmpl "istio.io/mixer/pkg/adapter/template"
 	{{range .TemplateModels}}
 		"{{.PackageImportPath}}"
@@ -52,7 +53,7 @@ import (
 )
 
 var (
-	SupportedTmplInfo = map[string]Info {
+	SupportedTmplInfo = map[string]template.Info {
 	{{range .TemplateModels}}
 		{{.GoPackageName}}.TemplateName: {
 			CtrCfg:  &{{.GoPackageName}}.InstanceParam{},
@@ -67,7 +68,7 @@ var (
 				_, ok := hndlr.({{.GoPackageName}}.{{.Name}}Handler)
 				return ok
 			},
-			InferType: func(cp proto.Message, tEvalFn TypeEvalFn) (proto.Message, error) {
+			InferType: func(cp proto.Message, tEvalFn template.TypeEvalFn) (proto.Message, error) {
 				var err error = nil
 				cpb := cp.(*{{.GoPackageName}}.InstanceParam)
 				infrdType := &{{.GoPackageName}}.Type{}
@@ -127,7 +128,7 @@ var (
 					for name, md := range castedInsts {
 						{{range .TemplateMessage.Fields}}
 							{{if isStringValueTypeMap .GoType}}
-								{{.GoName}}, err := evalAll(md.{{.GoName}}, attrs, mapper)
+								{{.GoName}}, err := template.EvalAll(md.{{.GoName}}, attrs, mapper)
 							{{else}}
 								{{.GoName}}, err := mapper.Eval(md.{{.GoName}}, attrs)
 							{{end}}
@@ -178,7 +179,7 @@ var (
 					for name, md := range castedInsts {
 						{{range .TemplateMessage.Fields}}
 							{{if isStringValueTypeMap .GoType}}
-								{{.GoName}}, err := evalAll(md.{{.GoName}}, attrs, mapper)
+								{{.GoName}}, err := template.EvalAll(md.{{.GoName}}, attrs, mapper)
 							{{else}}
 								{{.GoName}}, err := mapper.Eval(md.{{.GoName}}, attrs)
 							{{end}}
@@ -218,7 +219,7 @@ var (
 					castedInst := inst.(*{{.GoPackageName}}.InstanceParam)
 					{{range .TemplateMessage.Fields}}
 						{{if isStringValueTypeMap .GoType}}
-							{{.GoName}}, err := evalAll(castedInst.{{.GoName}}, attrs, mapper)
+							{{.GoName}}, err := template.EvalAll(castedInst.{{.GoName}}, attrs, mapper)
 						{{else}}
 							{{.GoName}}, err := mapper.Eval(castedInst.{{.GoName}}, attrs)
 						{{end}}
